fix: 🐛 dashboard remove
get rid of runtime error after dashboard is removed
