+++ src/modules/dashboards/saga/index.ts
@@ -2,5 +2,14 @@ import { saveDashboard } from './saveDashboard';
 import { cloneDashboard } from './cloneDashboard';
 import { viewDashboard } from './viewDashboard';
 import { resetDashboardFilters } from './resetDashboardFilters';
+import { deleteAccessKey } from './deleteAccessKey';
+import { deleteDashboard } from './deleteDashboard';
 
-export { viewDashboard, saveDashboard, cloneDashboard, resetDashboardFilters };
+export {
+  viewDashboard,
+  deleteDashboard,
+  deleteAccessKey,
+  saveDashboard,
+  cloneDashboard,
+  resetDashboardFilters,
+};
